MessagesViewModel: emit onUnreadChanged, drop Dock side-effects; fix stale UInt32 in chat test fake

+++ Packages/HeidrunModules/Tests/HeidrunChatTests/UserListViewModelTests.swift
@@ -314,7 +314,7 @@ private final class FakeUserListClient: HotlineClient, @unchecked Sendable {
     func startFolderDownload(at path: RemotePath, name: String) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
-    func startUpload(at path: RemotePath, name: String, size: UInt32, resume: Bool) async throws -> TransferHandle {
+    func startUpload(at path: RemotePath, name: String, size: UInt64, resume: Bool) async throws -> TransferHandle {
         TransferHandle(transferID: 0, totalSize: 0)
     }
     func startFolderUpload(at path: RemotePath, name: String, size: UInt64, itemCount: UInt16, resume: Bool) async throws -> TransferHandle {
